schemaを分解
