+++ app/repositories/run_date.py
@@ -2,7 +2,7 @@
 from sqlalchemy.orm import Session
 
 from app.models.run_date import RunDate
-from app.schemas import schemas
+from app.schemas.run_date import RunDateCreate
 
 
 class RunDateRepository:
@@ -12,7 +12,7 @@ class RunDateRepository:
     def get_run_dates(self, skip: int = 0, limit: int = 100):
         return self.db.query(RunDate).offset(skip).limit(limit).all()
 
-    def create_run_date(self, run_date: schemas.RunDateCreate):
+    def create_run_date(self, run_date: RunDateCreate):
         db_run_date = RunDate(
             task_id=run_date.task_id,
             date=run_date.date
